backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
openai>=1.0.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.0.0
openpyxl>=3.1.0
# PostgreSQL (users, sessions, messages + RAG)
asyncpg>=0.29.0
sentence-transformers>=3.0.0
pgvector>=0.3.0
# PageIndex (vectorless tree-based RAG)
pymupdf>=1.26.0
PyPDF2>=3.0.0
tiktoken>=0.11.0
pyyaml>=6.0.0
# Error tracking
sentry-sdk[fastapi]>=2.0.0
# Prometheus metrics
prometheus_client>=0.20.0
# Redis for rate limiting
redis>=5.0.0
